vllm-peteish-requirements.txt
accelerate==0.34.2
-e git+https://github.com/allenai/OLMo.git@fb659d3b8d1feed43ce72fcad23cc9d564977de1#egg=ai2_olmo
ai2-olmo-core==0.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
argcomplete==3.5.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
awscli==1.34.21
bitsandbytes==0.42.0
black==24.8.0
boto3==1.35.21
botocore==1.35.21
buildkite-test-collector==0.1.8
cached_path==1.6.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
clang-format==18.1.5
click==8.1.7
cloudpickle==3.0.0
codespell==2.3.0
colorama==0.4.6
compressed-tensors==0.4.0
contourpy==1.3.0
cupy-cuda12x==13.3.0
cycler==0.12.1
datamodel-code-generator==0.26.0
datasets==3.0.0
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docutils==0.16
einops==0.8.0
email_validator==2.2.0
exceptiongroup @ file:///croot/exceptiongroup_1706031385326/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastapi==0.115.0
fastrlock==0.8.2
filelock==3.13.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
genson==1.3.0
gguf==0.10.0
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
h11==0.14.0
hiredis==3.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.23.5
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflect==5.6.2
iniconfig==2.0.0
interegular==0.3.3
ipython @ file:///croot/ipython_1726064227985/work
isort==5.13.2
jedi @ file:///croot/jedi_1721058342488/work
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lark==1.2.2
lazy_loader==0.4
libnacl==2.1.0
librosa==0.10.2.post1
llvmlite==0.43.0
lm-format-enforcer==0.10.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mdurl==0.1.2
mistral_common==1.4.1
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.16
mypy==1.11.1
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==1.46.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
outlines==0.0.46
packaging==24.1
pandas==2.2.2
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
partial-json-parser==0.2.1.1.post4
pathspec==0.12.1
peft==0.12.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
proto-plus==1.24.0
protobuf==5.28.1
psutil==6.0.0
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py==1.11.0
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments @ file:///croot/pygments_1684279966437/work
pyparsing==3.1.4
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-forked==1.6.0
pytest-rerunfailures==14.0
pytest-shard==0.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.35.0
redis==5.0.8
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
rsa==4.7.2
ruff==0.1.5
s3transfer==0.10.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.0
sentencepiece==0.2.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.38.5
sympy==1.13.2
tensorizer==2.9.0
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.9
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.5
traitlets @ file:///croot/traitlets_1718227057033/work
transformers==4.44.2
transformers-stream-generator==0.0.5
triton==3.0.0
types-PyYAML==6.0.12.20240917
types-requests==2.32.0.20240914
types-setuptools==75.1.0.20240917
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
-e git+https://github.com/AkshitaB/vllm.git@0b8a08e4b963f2b84fb0f65533654d5b85e580e3#egg=vllm
vllm-flash-attn==2.6.1
watchfiles==0.24.0
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
websockets==13.0.1
xformers==0.0.27.post2
xxhash==3.5.0
yapf==0.32.0
yarl==1.11.1
zipp==3.20.2
